Fix ssl server, client certificate checks with cert store
When a certificate store was enabled in the SSL context, the Trust
Manager used was doing nothing, thus never checking the validity
of the client certificate.
This patch makes sure the SSLContext's Trust Manager is used instead of
the dummy Store one. Hopefully This TM knows how to use the store content
to validate certs.

Signed-off-by: Brice Figureau <brice@daysofwonder.com>