Fix all openssl tests
Some openssl tests where failing for various reasons.
The most notable one was the use of ByteList.plain
to create strings. This seems, at least on my mac to
produce strange errors.

The other error was related to the possibility when
using Purpose that nsCertType could not be an extension
of the current cert.

Signed-off-by: Brice Figureau <brice@daysofwonder.com>